+++ src/mastodon_api/accounts/types.rs
@@ -305,12 +305,6 @@ impl AccountUpdateData {
     }
 }
 
-#[derive(Deserialize)]
-pub struct MoveFollowersRequest {
-    pub from_actor_id: String,
-    pub followers_csv: String,
-}
-
 #[derive(Serialize, Deserialize)]
 #[serde(tag = "type", rename_all = "kebab-case")]
 pub enum ActivityParams {