+++ src/mastodon_api/settings/types.rs
@@ -4,3 +4,9 @@ use serde::Deserialize;
 pub struct PasswordChangeRequest {
     pub new_password: String,
 }
+
+#[derive(Deserialize)]
+pub struct MoveFollowersRequest {
+    pub from_actor_id: String,
+    pub followers_csv: String,
+}